Relocate scanner to blob package
